RISC-V: Default to medany on 64-bit ELF targets
Many users of RISC-V toolchains are interested in embedded systems,
where riscv*-*-elf is the standard tuple.  These systems tend to have
memory distributed throughout the address space.  For 32-bit systems the
default medlow code model is fine as it covers the entire address space,
but for 64-bit systems the medany code model is commonly used because
there tend to be memories outside the range of medlow.

As a result, pretty much every toolchain vendor is setting some sort of
CFLAGS_FOR_TARGET="... -mcmodel=medany ..." when building their
toolchians.  This causes two issues:

* The vendors that forget end up with broken toolchians -- specifically
  this recently came up in the context of the Zephyr SDK.
* The vendors that remember still end up with medany builds of the
  32-bit libraries.  There's a small performance penalty here, but also
  a possible usability headache related to the PC-relative relocation
  alignment issue.

This patch changes the default code model to medany on riscv*-*-elf
toolchains when generating code for 64-bit targets.  Another option
would be to introduce a mechanism into t-elf-multilib (presumably via
multilib-generator) to allow users to select the code model for the
64-bit libraries, but this seems like a more user-friendly option to me.

The obvious issue here is that this patch changes the default behavior
of riscv*-*-elf-gcc, but given the change matches what most users are
asking for I think it's sane.

2019-08-07  Palmer Dabbelt  <palmer@sifive.com>

        * config/riscv/riscv.h (TARGET_DEFAULT_CMODEL): Remove
        definition.
        * config/riscv/elf.h (TARGET_DEFAULT_CMODEL): Moved from
        riscv.h.
        * config/riscv/linux.h (TARGET_DEFAULT_CMODEL): Likewise.
        * config/riscv/riscv-opts.h (enum riscv_cmodel): New CM_DEFAULT
        entry.
        * config/riscv/riscv.c (riscv_option_override): Select a code
        model based on the ISA.