Add TARGET_ARG_EXTENEDED_ON_STACK
We've had 82106 sitting around for a while, Jeff found it when digging
through some old bugs.  Jim had posted a WIP patch in there and I've
basically just took his patch and plumbed it through to a target hook
that checks for the misaligned case to gate

Since Jim wrote the actual code I'm not entirely sure I understand
what's going on, but just from reading through it the old comments seem
to be invalidated by some of this new code.  So I need to go read this
and understand it enough to write up some comments that makes sense, but
that might take a bit.

	PR 82106

gcc/ChangeLog:

	* config/riscv/riscv.cc (struct riscv_arg_info): Add ap_offset
	field.
	(riscv_get_arg_info): Populate ap_offset.
	(riscv_arg_extended_on_stack): New hook.
	(TARGET_ARG_EXTENDED_ON_STACK): Likewise.
	* doc/tm.texi: Regenerate.
	* doc/tm.texi.in: Add TARGET_ARG_EXTENDED_ON_STACK hook.
	* function.cc (struct assign_parm_data_one): Check
	TARGET_ARG_EXTENDED_ON_STACK.
	(assign_parm_find_entry_rtl): Likewise.
	(assign_parm_is_stack_parm): Likewise.
	(assign_parm_adjust_entry_rtl): Likewise.
	(assign_parm_setup_reg): Likewise.
	(assign_parm_setup_stack): Likewise.
	* target.def: Define TARGET_ARG_EXTENDED_ON_STACK.
	* targhooks.cc (hook_int_CUMULATIVE_ARGS_arg_info_1): New hook
	default.
	* targhooks.h: Likewise

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/pr82106.c: New test.